Added a "normal" frame option.
This is due to the several requests by users
to not have the frame that puts the phone
number in the corner of the screen.  This also
means that you can only have one extra item
in that mode. i.e. misc1 or misc2 but not both
due to the phone number taking extra space
between the pictures.

Signed-off-by: Gabe <gabenbecca@gmail.com>